src/gateway/requirements.txt
asyncpg>=0.24.0
fastapi>=0.70.0
loguru>=0.5.3
aiosql>=3.3.1
uvicorn>=0.15.0
hypercorn>=0.12.0
pydantic>=1.8.2
PyJWT>=2.3.0
email-validator>=1.1.3
passlib>=1.7.4
bcrypt>=3.2.0
httpx>=0.20.0
orjson>=3.6.4